move getAllUsedAddresses back to computed
